app/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
gunicorn==20.1.0
phonenumbers==8.12.45
psycopg2-binary==2.9.3
sqlparse==0.4.2
